# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/licenses/{id}": {
      "get": {
        "description": "Public license detail including type, business, and documents with presigned URLs (10-min TTL).",
        "operationId": "LicenseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "License uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The license with related data."
          },
          "404": {
            "description": "License not found."
          }
        },
        "summary": "Get a license by id",
        "tags": [
          "Licenses"
        ]
      }
    },
    "/api/licenses/{id}/qr-verify": {
      "get": {
        "description": "Minimal license payload for QR scanning. Rate limited to 60/min/IP.",
        "operationId": "LicenseController_verifyQr",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "License uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Minimal license verification payload."
          },
          "404": {
            "description": "License not found."
          }
        },
        "summary": "QR verification of a license",
        "tags": [
          "Licenses"
        ]
      }
    },
...
    "/api/my/notifications": {
      "get": {
        "description": "Paginated, newest first. Returns `{ data, meta }`.",
        "operationId": "NotificationController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paginated notifications."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List my notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/read-all": {
      "patch": {
        "description": "Marks every unread notification for the current user as read.",
        "operationId": "NotificationController_readAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Count of updated notifications."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mark all my notifications read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "description": "Marks a single notification (owned by the user) as read.",
        "operationId": "NotificationController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated notification."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mark one notification read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "description": "Admin sees all users; supervisor sees only their own agency. Filter by role / zone / status / free-text query.",
        "operationId": "UserController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Matching users."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Creates a SUPERVISOR/INSPECTOR account. A temp password is generated for self-login users (mustChangePassword=true).",
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created user (with temp password)."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a user (admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/roles": {
      "patch": {
        "description": "Replaces the user’s role set. Admin-only by locked decision.",
        "operationId": "UserController_roles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Set user roles (admin only — D1)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/agency": {
      "patch": {
        "description": "Admin: any user. Supervisor: only within their own agency (else 403).",
        "operationId": "UserController_agency",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user."
          },
          "403": {
            "description": "Cross-agency change by a supervisor."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Set user agency",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/zones": {
      "post": {
        "description": "Replaces the user’s zone assignments. Supervisors may only assign their own zones.",
        "operationId": "UserController_zones",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZonesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The updated zone assignments."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace a user’s zones",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/suspend": {
      "patch": {
        "description": "Sets isActive=false and revokes all of the user’s sessions.",
        "operationId": "UserController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The suspended user."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Suspend a user (admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "delete": {
        "description": "Soft-deletes the user and revokes all sessions.",
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a user (admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/zones": {
      "get": {
        "description": "Admin sees all zones; supervisor sees their own zones.",
        "operationId": "ZoneController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Zones."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List zones",
        "tags": [
          "Zones"
        ]
      },
      "post": {
        "description": "Creates a new zone with a GeoJSON boundary.",
        "operationId": "ZoneController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created zone."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a zone (admin)",
        "tags": [
          "Zones"
        ]
      }
    },
    "/api/zones/{id}": {
      "put": {
        "description": "Updates zone fields (name, province, boundary, active flag).",
        "operationId": "ZoneController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Zone uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated zone."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a zone (admin)",
        "tags": [
          "Zones"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-License Verification Platform API",
    "description": "REST API for the Thai e-licensing platform that plugs into the ทางรัฐ\n(Tang Rat) super app. Serves the public app and the admin portal.\n\n**Authentication.** Most endpoints require a Bearer access token\n(RS256 JWT, 15-min TTL). Obtain one from `POST /api/auth/register` or\n`POST /api/auth/login` (PUBLIC), `POST /api/auth/tang-rat`\n(PUBLIC/INSPECTOR/SUPERVISOR), or `POST /api/auth/self` (ADMIN + TOTP).\nClick **Authorize** and paste the `accessToken` to call protected routes.\nSee `docs/AUTHENTICATION.md` for a plain-English walkthrough.\n\n**Scope.** INSPECTOR/SUPERVISOR responses are filtered server-side by the\ncaller's zones and agency; clients cannot widen scope via query params.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Paste the access token returned by an /auth login route."
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateRolesDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAgencyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateZonesDto": {
        "type": "object",
        "properties": {}
      },
      "CreateZoneDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateZoneDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}